fix: move main.ts to root directory for proper compilation output
Moved `src/core/main.ts` to `src/main.ts` so that it correctly compiles to `scripts/main.js` which is what the `manifest.json` expects. Also updated `tsup.config.ts` and relative imports.
