+++ src/features/essentials/commands/reload.ts
@@ -1,6 +1,6 @@
 import { reloadConfig } from '@core/configManager.js';
 import { errorLog } from '@core/logger.js';
-import { updateAllPlayerRanks } from '@core/main.js';
+import { updateAllPlayerRanks } from '../../../main.js';
 import { sendMessage } from '@core/messaging.js';
 
 import { CustomCommand } from '@commands/commandManager.js';
